testing and modularizing new components
