+++ package.json
@@ -42,6 +42,7 @@
     "babel-preset-es2015": "^6.24.1",
     "can-connect": "^1.5.14",
     "can-fixture": "^1.2.2",
+    "can-vdom": "^3.2.5",
     "eslint": "^4.14.0",
     "http-server": "^0.10.0",
     "jest": "^22.0.4",
